% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlackOutDates.R
\docType{data}
\name{BlackOutDates}
\alias{BlackOutDates}
\title{BlackOutDates}
\format{
A data frame with the following columns:
\describe{
\item{PassType}{The ski season the pass type was for.}
\item{Date}{The passtype}
}
}
\usage{
BlackOutDates
}
\description{
Because certain passes have blackout dates where the customer is not allowed
to ski, we need some way of keeping track of what days are allowed. This table
documents which days are not allowed.
}
\keyword{datasets}
