% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassTypes.R
\docType{data}
\name{PassTypes}
\alias{PassTypes}
\title{PassTypes}
\format{
A data frame with the following columns:
\describe{
\item{Season}{The ski season the pass type was for.}
\item{PassType}{The passtype}
}
}
\usage{
PassTypes
}
\description{
This is the list of passes that are available to be sold in each season.
}
\keyword{datasets}
