% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatrolIssues.R
\docType{data}
\name{PatrolIssues}
\alias{PatrolIssues}
\title{PatrolIssues}
\format{
A data frame with the following columns:
\describe{
\item{PersonID}{A unique identifier identifying a person within the database.}
\item{Date}{The date of the incident.}
\item{Issue}{A description of the incident.}
}
}
\usage{
PatrolIssues
}
\description{
This table is contains documentation for every patrol related interaction
that we want to document.
}
\keyword{datasets}
