% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkiPasses_sqlite.R
\name{SkiPasses_sqlite}
\alias{SkiPasses_sqlite}
\title{Create database connection for Ski Passes Database}
\usage{
SkiPasses_sqlite(path = ":memory:")
}
\arguments{
\item{path}{The sqlite database file to be created. If missing, then the database
is stored "in memory".}
}
\description{
This function pulls the example Ski Pass data into a sqlite database. If
the user has specified a `path`, then we'll save the database to a file, which
can be subsequently updated.
}
\details{
There are five tables we are interested in:

`Customers` is contains personal information about each customer.

`Passes` contains information about each season pass. In particular a single
         customer might have season passes in different years. Furthermore

`PassTypes` This table contains all the different pass types sold for each year.

`BlackOutDates` Certain passes have black-out dates where the customer is not
                allowed to ski. These are typically on busy vacation weekends.
                Also the Weekday pass has *every* weekend blacked out.

`PatrolIssues` This table documents any time a customer has their pass pulled
               due to hooligan behavior such as ducking boundary ropes and
               building jumps in dangerous locations.
}
